requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
groq>=0.11.0
sentence-transformers>=3.0.0
scikit-learn>=1.5.0
numpy>=1.26.0
PyMuPDF>=1.24.0
pydantic>=2.7.0
rich>=13.7.0
python-dotenv>=1.0.0
flask>=3.0.0
